+++ src/lib.rs
@@ -16,7 +16,6 @@ extern crate urlencoded;
 
 use emailaddress::EmailAddress;
 use hyper::client::Client as HttpClient;
-use iron::headers::ContentType;
 use iron::modifiers;
 use iron::prelude::{IronResult, Response};
 use iron::status;
@@ -24,7 +23,6 @@ use iron::Url;
 use lettre::transport::EmailTransport;
 use openssl::crypto::hash;
 use openssl::crypto::pkey::PKey;
-use serde_json::builder::ObjectBuilder;
 use serde_json::de::from_reader;
 use serde_json::value::Value;
 use rand::{OsRng, Rng};
@@ -34,7 +32,6 @@ use std::collections::BTreeMap;
 use std::fs::File;
 use std::io::{BufReader, Write};
 use std::iter::Iterator;
-use time::now_utc;
 use url::percent_encoding::{utf8_percent_encode, QUERY_ENCODE_SET};
 use urlencoded::QueryMap;
 
@@ -231,57 +228,4 @@ fn create_jwt(key: &PKey, header: &str, payload: &str) -> String {
     input.push(b'.');
     input.extend(sig.to_base64(base64::URL_SAFE).into_bytes());
     String::from_utf8(input).unwrap()
-}
-
-
-/// HTML template used to have the user agent POST the identity token built
-/// by the daemon instance to the RP's `redirect_uri`.
-const FORWARD_TEMPLATE: &'static str = r#"<!DOCTYPE html>
-<html>
-  <head>
-    <title>Let's Auth</title>
-    <script>
-      document.addEventListener('DOMContentLoaded', function() {
-        document.getElementById('form').submit();
-      });
-    </script>
-  </head>
-  <body>
-    <form id="form" action="{{ return_url }}" method="post">
-      <input type="hidden" name="id_token" value="{{ jwt }}">
-    </form>
-  </body>
-</html>"#;
-
-
-/// Iron handler for sending an identity token to the Relying Party.
-///
-/// Builds the JWT header and payload JSON data and signs it with the
-/// configured private RSA key. Then uses `FORWARD_TEMPLATE` to embed the token
-/// in a form that's POSTed to the RP's `redirect_uri` as soon as the page
-/// is loaded.
-fn send_jwt_response(app: &AppConfig, email: &str, origin: &str, redirect: &str) -> IronResult<Response> {
-    let now = now_utc().to_timespec().sec;
-    let payload = ObjectBuilder::new()
-        .insert("aud", origin)
-        .insert("email", email)
-        .insert("email_verified", email)
-        .insert("exp", now + app.token_validity)
-        .insert("iat", now)
-        .insert("iss", &app.base_url)
-        .insert("sub", email)
-        .unwrap();
-    let headers = ObjectBuilder::new()
-        .insert("kid", "base")
-        .insert("alg", "RS256")
-        .unwrap();
-    let jwt = create_jwt(&app.priv_key,
-                         &serde_json::to_string(&headers).unwrap(),
-                         &serde_json::to_string(&payload).unwrap());
-
-    let html = FORWARD_TEMPLATE.replace("{{ return_url }}", redirect)
-        .replace("{{ jwt }}", &jwt);
-    let mut rsp = Response::with((status::Ok, html));
-    rsp.headers.set(ContentType::html());
-    Ok(rsp)
 }
